requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
python-multipart
sqlalchemy
python-jose[cryptography]
bcrypt
psycopg2-binary
alembic
pydantic
pydantic-settings